7. het - Adatkezeles II - Gyakorlo feladat UPDATE ES JAVITASOK

+++ hr/src/main/java/hu/webuni/hr/vargyasb/service/AbstractEmployee.java
@@ -11,9 +11,11 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.domain.Specification;
+import org.springframework.http.HttpStatus;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.ObjectUtils;
 import org.springframework.util.StringUtils;
+import org.springframework.web.server.ResponseStatusException;
 
 import hu.webuni.hr.vargyasb.model.Company;
 import hu.webuni.hr.vargyasb.model.Employee;
@@ -25,57 +27,47 @@ public abstract class AbstractEmployee implements EmployeeService {
 
 	@Autowired
 	EmployeeRepository employeeRepository;
-	
+
 	@Autowired
 	PositionRepository positionRepository;
-	
+
 	@Transactional
 	public Employee save(Employee employee) {
-		Position position = employee.getPosition();
-		if (position != null) {
-			String positionName = position.getName();
-			if (!ObjectUtils.isEmpty(positionName)) {
-				Position positionInDb = null;
-				Optional<Position> foundPosition = positionRepository.findByName(positionName);
-				if (foundPosition.isPresent()) {
-					positionInDb = foundPosition.get();
-				} else {
-					positionInDb = positionRepository.save(position);
-				}
-				employee.setPosition(positionInDb);
-			} else {
-				employee.setPosition(null);
-			}
-		}
+		savePositionForEmployee(employee);
 		return employeeRepository.save(employee);
 	}
-	
-//	@Transactional
-//	public Employee update(Employee employee) {
-//		if (employeeRepository.existsById(employee.getId())) {
-//			return employeeRepository.save(employee);
-//		} else {
-//			throw new NoSuchElementException();
-//		}
-//	}
-	
+
+	@Transactional
+	public Employee update(Employee employee) {
+		if (employeeRepository.existsById(employee.getId())) {
+			savePositionForEmployee(employee);
+			return employeeRepository.save(employee);
+		} else {
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+		}
+	}
+
 	public Page<Employee> findAll(Pageable pageable) {
 		return employeeRepository.findAll(pageable);
 	}
-	
+
 	public Optional<Employee> findById(Long id) {
 		return employeeRepository.findById(id);
 	}
-	
+
 	@Transactional
 	public void delete(Long id) {
-		employeeRepository.deleteById(id);
+		if (employeeRepository.existsById(id)) {
+			employeeRepository.deleteById(id);
+		} else {
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+		}
 	}
-	
+
 	public List<Employee> getEmployeesWhoseSalaryIsGreaterThan(int salary, Integer pageNr, Integer pageSize) {
 		Pageable paging = PageRequest.of(pageNr, pageSize);
 		Page<Employee> pagedResult = employeeRepository.findBySalaryGreaterThan(salary, paging);
-		
+
 //		return employeeRepository.findBySalaryGreaterThan(salary, paging);
 		if (pagedResult.hasContent()) {
 			return pagedResult.getContent();
@@ -95,39 +87,59 @@ public abstract class AbstractEmployee implements EmployeeService {
 	public List<Employee> findByStartOfEmploymentBetween(LocalDateTime from, LocalDateTime to) {
 		return employeeRepository.findByStartOfEmploymentBetween(from, to);
 	}
-	
+
 	public List<Employee> findEmployeesByExample(Employee example) {
 		Long id = example.getId();
-		if(id == null)
+		if (id == null)
 			id = (long) 0;
 		String name = example.getName();
 		String positionName = null;
 		Position position = example.getPosition();
-		if(position != null)
+		if (position != null)
 			positionName = position.getName();
 		int salary = example.getSalary();
 		LocalDateTime startOfEmployment = example.getStartOfEmployment();
 		String companyName = null;
 		Company company = example.getCompany();
-		if(company != null)
+		if (company != null)
 			companyName = company.getName();
-		
+
 		Specification<Employee> spec = Specification.where(null);
-		
-		if(id > 0)
+
+		if (id > 0)
 			spec = spec.and(EmployeeSpecifications.hasId(id));
-		if(StringUtils.hasText(name))
+		if (StringUtils.hasText(name))
 			spec = spec.and(EmployeeSpecifications.hasName(name));
-		if(StringUtils.hasText(positionName))
+		if (StringUtils.hasText(positionName))
 			spec = spec.and(EmployeeSpecifications.hasPosition(positionName));
-		if(salary > 0)
+		if (salary > 0)
 			spec = spec.and(EmployeeSpecifications.hasSalary(salary));
-		if(startOfEmployment != null)
+		if (startOfEmployment != null)
 			spec = spec.and(EmployeeSpecifications.hasstartOfEmployment(startOfEmployment));
-		if(StringUtils.hasText(companyName))
+		if (StringUtils.hasText(companyName))
 			spec = spec.and(EmployeeSpecifications.hasCompany(companyName));
-			
+
 		return employeeRepository.findAll(spec, Sort.by("id"));
 	}
 	
+	@Transactional
+	private void savePositionForEmployee(Employee employee) {
+		Position position = employee.getPosition();
+		if (position != null) {
+			String positionName = position.getName();
+			if (!ObjectUtils.isEmpty(positionName)) {
+				Position positionInDb = null;
+				Optional<Position> foundPosition = positionRepository.findByName(positionName);
+				if (foundPosition.isPresent()) {
+					positionInDb = foundPosition.get();
+				} else {
+					positionInDb = positionRepository.save(position);
+				}
+				employee.setPosition(positionInDb);
+			} else {
+				employee.setPosition(null);
+			}
+		}
+	}
+
 }
